+++ lib/Report/Generator/ConsoleTableReportGenerator.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace PhpBench\ReportGenerator;
+namespace PhpBench\Report\Generator;
 
 use Symfony\Component\Console\Helper\Table;
 use Symfony\Component\Console\Output\OutputInterface;
